Add Python bindings for the fs, render, net, and ui namespaces
 ! This commit is way too big and I'm sorry but I'm too exhausted to properly
   split it up into manageable chunks :/
 ! On the bright side it's a huge step towards full Python usability
 - Removed ObjUIElement::set_is_visible() and ::set_focus() in favor of
   directly calling ::set_data() on such Instances
 * Moved all primitive drawing functions to the render namespace
 * Moved texture_before, texture_after, transition_type, transition_speed, and
   transition_custom_func from the EngineState struct to the hidden
   bee::render::internal namespace
 + Changed Instance::get_relation() to return the new E_RELATION enum
 + Added functions in bee/render/transition.cpp to manage the hidden Textures
 * Changed Variant to store longs instead of ints
 - Removed the ability to store floats in Variant in favor of doubles
 * Changed the associated SerialData methods to match Variant storage
 * Renamed E_DATA_TYPE::INT to ::INTEGER, and ::FLOAT to ::FLOATING
 - Removed E_DATA_TYPE::DOUBLE in favor of ::FLOATING
 - Removed the Variant access macro _d()
 * Moved console UI initialization inside bee::loop() to allow multi-phase
   loading for use with loading level maps
 + Added bee::get_fps_stable()
 * Moved compute_check_quit() from bee/render/transition.cpp to bee/engine.cpp
 + Added the E_FS_ROOT_TYPE enum to describe how fs paths should be mapped
 + Added the mapname member variable to the FilePath class
 * Changed the internal filesystem mapping to use a vector of FilePaths to
   represent individual files from multiple sources
 + Added documentation comments to the functions in bee/fs/fs.cpp
 * Rewrote fs::scan_files() to handle roots and archives
 - Removed /maps from the default filesystem mapping
 + Added stubs for fs::scan_archive() and ::load_level()
 + Added a Python binding in bee/fs/python.cpp which returns a callable that
   can be used to get the contents of a FilePath more lazily
 + Added Python module bindings for the fs, render, net, and ui namespaces
 * Replaced the Python enum for E_RGB with a dict of RGBA tuples
 + Added Python bindings for the missing Instance methods
 + Added Python rich compare and hash functions for the Instance class based on
   the Instance id
 + Added Python bindings for the missing kb module functions
 * Fixed python::run_file() to use the filesystem
 * Changed PythonScriptInterface::run_string() to take a filename argument
 * Changed the <Resource>_from() helper functions to safely handle nullptr
 * Fixed parsing of vector doubles in the Python bindings
 * Changed ObjUIElement to use chars for it's color data instead of ints
 * Rewrote ObjControl and ObjBee in Python
 - Removed the old C++ versions of RmTest, ObjControl, and ObjBee

+++ bee/init/gameoptions.cpp
@@ -128,7 +128,7 @@ namespace bee {
 				return 0;
 			}}},
 
-			{"renderer_type", {"renderer_type", Variant(static_cast<int>(E_RENDERER::OPENGL3)), [] (GameOption* self, Variant value) {
+			{"renderer_type", {"renderer_type", Variant(static_cast<long>(E_RENDERER::OPENGL3)), [] (GameOption* self, Variant value) {
 				self->value = value;
 
 				if (engine->renderer->context != nullptr) {
